Crittercism Module and library additions
