+++ app/config.json
@@ -6,8 +6,8 @@
       "remoteLevel": "E",
       "fileLevel": "D",
       "crittercism": false,
-      "crittercismIOSAppId": "54a3119851de5e9f042ec6d5",
-      "crittercismAndroidAppId": "54a30d4651de5e9f042ec6d2"
+      "crittercismIOSAppId": "55184773bdf36ccd74eae9d8",
+      "crittercismAndroidAppId": "drlASD5bxwrFXSHXpbWScR5d0MXHUEu9"
     }
   },
   "env:development": {},
