+++ modules/android/crittercism/1.0.1/LICENSE
@@ -1,0 +1,1 @@
+TODO: place your license here and we'll include it in the module distribution
