+++ modules/android/crittercism/1.0.1/manifest
@@ -1,0 +1,17 @@
+#
+# this is your module manifest and used by Titanium
+# during compilation, packaging, distribution, etc.
+#
+version: 1.0.1
+apiversion: 2
+description: CrittercismModule
+author: Mike Fogg
+license: None
+copyright: Copyright (c) 2014
+
+# these should not be edited
+name: CrittercismModule
+moduleid: crittercism
+guid: 0796ee2e-df24-4604-b8cf-a9cb17e87f3d
+platform: android
+minsdk: 3.0.0
